fix(split-reveal): share hooks and stabilize bootstrap lifecycle
Move scroll lock and reduced-motion hooks to hooks/, consolidate on
useLockBody, and skip re-bootstrap until the first run completes so
Images task registration does not abort preload.
