+++ hooks/use-lock-body.ts
@@ -1,10 +1,47 @@
-import * as React from "react";
+"use client";
 
-// @see https://usehooks.com/useLockBodyScroll.
-export function useLockBody() {
-  React.useLayoutEffect((): (() => void) => {
-    const originalStyle: string = window.getComputedStyle(document.body).overflow;
-    document.body.style.overflow = "hidden";
-    return () => (document.body.style.overflow = originalStyle);
-  }, []);
+import { useEffect } from "react";
+
+export function useLockBody(active = true) {
+  useEffect(() => {
+    if (!active) {
+      return;
+    }
+
+    const html = document.documentElement;
+    const body = document.body;
+    const scrollY = window.scrollY;
+
+    const previous = {
+      htmlOverflow: html.style.overflow,
+      bodyOverflow: body.style.overflow,
+      bodyPosition: body.style.position,
+      bodyTop: body.style.top,
+      bodyLeft: body.style.left,
+      bodyRight: body.style.right,
+      bodyWidth: body.style.width,
+      bodyTouchAction: body.style.touchAction,
+    };
+
+    html.style.overflow = "hidden";
+    body.style.overflow = "hidden";
+    body.style.position = "fixed";
+    body.style.top = `-${scrollY}px`;
+    body.style.left = "0";
+    body.style.right = "0";
+    body.style.width = "100%";
+    body.style.touchAction = "none";
+
+    return () => {
+      html.style.overflow = previous.htmlOverflow;
+      body.style.overflow = previous.bodyOverflow;
+      body.style.position = previous.bodyPosition;
+      body.style.top = previous.bodyTop;
+      body.style.left = previous.bodyLeft;
+      body.style.right = previous.bodyRight;
+      body.style.width = previous.bodyWidth;
+      body.style.touchAction = previous.bodyTouchAction;
+      window.scrollTo(0, scrollY);
+    };
+  }, [active]);
 }
